requirements.txt
gradio
pillow
soundfile
imageio
apng
pydub
ffmpeg